Moved TableStyle and MemoryStyle to wasmer-types (and compiler-single pass do not depend from vm anymore)
