+++ lib/types/src/lib.rs
@@ -59,9 +59,11 @@ mod extern_ref;
 mod features;
 mod indexes;
 mod initializers;
+mod memory;
 mod memory_view;
 mod module;
 mod native;
+mod table;
 mod trapcode;
 mod types;
 mod units;
@@ -95,8 +97,9 @@ pub use types::{
 #[cfg(feature = "enable-rkyv")]
 pub use archives::ArchivableIndexMap;
 
+pub use crate::memory::{MemoryError, MemoryStyle};
+pub use crate::table::TableStyle;
 pub use crate::trapcode::TrapCode;
-
 pub use crate::vmoffsets::{TargetSharedSignatureIndex, VMBuiltinFunctionIndex, VMOffsets};
 
 /// Version number of this crate.
